Major restructuring: resectioning and change of punchline
Previously, I had three bullet points for a single ''why do I care slides''.

Now, ``Why do I care'' becomes a section in which I explain why we might be
inefficiently ignoring advances in version control technology that people in
Tech have been using for long.

So I now show the "bad project screenshot" earlier, as an example of how we do
stuff inefficiently, instead of showing it in the -- now deleted -- 'Introducing
Git' section.

I also remove the "Git isn't" slide, which was a bit useless to begin with.